Increase Readability
There was a lot of code that was uneccessary and could have been made
more concise.

Increased readability would help people understand the code better

Removed irrelevant code in Duke, Made abstraction level in Parse method
the same, changed structure of switch statement
